requirements.txt
aiohttp==3.9.3
aiosignal==1.2.0
attrs==23.1.0
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
Flask==2.2.5
frozenlist==1.4.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.3
MarkupSafe==2.1.3
multidict==6.0.4
pip==23.3.1
requests==2.31.0
setuptools==68.2.2
urllib3==2.1.0
Werkzeug==2.3.8
wheel==0.41.2
yarl==1.9.3

openai~=1.9.0
python-dotenv~=0.21.0